Update message on successful external screening
This message is shown when an external source has been successfully
queried and returned an empty set of search results. Telling the user
to manually verify reduces confidence in the automatic screening; saying
simply that no records were matched is precise and lets the user decide
what that means in context.

Issue #289 Use LEIE API to automatically screen providers