Indicate if auto screen run on enrollment review
Change the link text for the automatic screening results to be either
"View results" if the automatic screening was run, or "Not performed" if
it was not. Hopefully this will somewhat clarify the difference between
"not performed" and "performed with a negative result", both of which
were previously indicated by an unchecked box in the "Verified" column.

Issue #354 Redesign Enrollment Review page