Default to "Verified" when auto screen not run
It is hard to see the progress we've made in enabling the LEIE automatic
screening. The significance of an absent check mark, indicating we found
a matching NPI in the LEIE and the enrollment should be rejected, is
lost in a sea of empty checkboxes. Mark results that have had no
automatic screening run as "Verified" to highlight this contrast.