Remove unused dependency on asm
One of the libraries that was originally vendored was asm[1]. We
intentionally moved it over to gradle without examining its use or
purpose so that we could use gradle instead of ant, with the promise
to look at it more closely later.

We have no direct dependency on any of the classes included in the asm
jar. They are all in a subpackage of `org.objectweb`, which we never
import. The vendored libraries we have from TopCoder do not depend on
it, either, according to Tattletale[2]. Finally, it turns out that
WildFly 10.1.0.Final bundles a newer version of asm, so even if we did
need it, it's provided at runtime.

Issue #16 Manage sets of dependencies via Gradle or another tool

[1] https://en.wikipedia.org/wiki/ObjectWeb_ASM
[2] http://tattletale.jboss.org/