Compile model to Java 1.6 for Guvnor
In order to use the guided editor, the Guvnor server needs a jar of the
model classes. Since the version of Drools and Guvnor we're using
requires an old version of JBoss AS, which in turn requires an old
version of Java, we must compile the jar at a language level it can
understand.