Reformat entity XML definitions
The model classes are generated from XML. In preparation for adding new
entities to the XML, reformat the primary XML file to use two-space
indentation (instead of a mix of tabs and spaces) and collapse empty
elements (so from `<a></a>` to `<a/>`). I used `xmllint` to do the
reformating:

    xmllint --format --noblanks

I also tweaked the root element to wrap across multiple lines.