Fix calls to `assertThat` to test what we expect
AssertJ[1] provides a fluent API for assertions and more-useful error
messages than the built-in Java `assert()` call, but some of its fluent
API methods overlap with methods on `String`s, which has led to some
mistaken calls. `assertThat(...)` returns an object; if no methods on
that object are called, nothing is asserted.

Move the method calls that we expect to assert outside of the
`assertThat()`.

[1] https://joel-costigliola.github.io/assertj/