Extract "Create New Enrollment" button to tag
The HTML of the "Create New Enrollment" button was very similar in the
five places it occurs, differing only in the CSS classes applied.
Extract a new custom tag and call it from each of those places, passing
in the CSS class(es) as an attribute.

Note that the attribute name `class` is reserved, which is why I named
the attribute `cssClass`.

Issue #333 Make the "Create New Enrollment" Button Testable