Add id attribute to Create Enrollment btn for test
Give the "Create New Enrollment" button a unique ID so that Selenium can
more reliably find it to click on it. The button has different CSS paths
depending on if the current user has already created a new enrollment;
we had only been testing with a user that did have enrollments created,
and so the tests were failing on a fresh database.

Issue #333 Make the "Create New Enrollment" Button Testable