Remove commons-logging runtime dependency
We do not use the Apache Commons Logging component[1] directly. Spring
uses it, and depends on it, so we still get it transitively. The only
direct reference to it in our repository, in a Java file that is bundled
as part of a JavaScript library, and that Java file isn't included in
our build.

Issue #16 Manage sets of dependencies via Gradle or another tool

[1] https://commons.apache.org/proper/commons-logging/