Add contextual explanation of NPI abbreviation
This commit reuses the PSM's modal-related infrastructure to provide
users with an explanation of the NPI, available by hovering over a
question mark in the form. This requires a modification of the CSS;
"definitionsModal" ought to be a class, not an ID. This also replaces
an old, flat NPI explanation with a richer (HTML) one.

Thanks to Cecilia Donnelly for pair-programming.

Fixes #265.