Reformat our JavaScript using JSCS
Run JSCS with the rules introduced in the previous commit on our
application's JavaScript (but not our vendored JavaScript libraries) to
clean up whitespace and formatting. Also manually touch up a few issues
JSCS wouldn't fix, including tabs in comments and multiple newlines in a
row.

These changes should be whitespace only, and have no effect on the
functionality of the JavaScript. This is just to make changing them
easier in the future.