Add Review button to pending enrollment admin view
I find that I frequently want to be able to review an enrollment
directly from the enrollment details page. Previously, the only way was
to go to the enrollment list, go to the pending tab, find the enrollment
in the list, and click "Review". Add a button to pending enrollments
that allow the admin to go directly to the review page.