Extract helper to add named address to model
Note that we now bind `attentionTo` for some addresses which we did not
do previously; the bound value is `null`, and the page template never
refers to it. Binding a null value is not an error, as demonstrated by
the fact that we were binding the reimbursement county, which is also
never referred to and defaults to null.

Refactoring in preparation to alter behavior around address lines.

Issue #8 Reduce code duplication
Issue #292 Label all required fields