Remove indices, search page from table of contents
The search page is available via the sidebar, and we
currently do not use the indices. If we decide to use
http://www.sphinx-doc.org/en/stable/markup/misc.html#index-generating-markup
at some future time, we can add these links back to
the ToC. This commit removes them to reduce clutter in
the user help Table of Contents and in the exported
PDF and ePub files.