Prioritize first line of address fields
There was a Drools validation rule that required address line 2 to be
present if either line 1 or 2 were present. This mismatch with how we
typically use address fields in the United States was previously being
"fixed" (or perhaps, more accurately, worked around) by testing (in Java
code) if the first line was empty, and if so switch lines 1 and 2. This
pattern was repeated in every place along the boundry between JSPs and
controllers that dealt with addresses.

One of the places, dealing with the home residence address of the
individual ownership interest, did not properly switch the two address
lines, leading to the second address line being marked as in error when
neither are filled, rather than the first address line.

I found the logic around switching the address lines very confusing, and
so rather than fix the symptom, fix the underlying problem: update the
Drools rule to prioritize address line one rather than address line two,
and remove all the complexity around switching address lines everywhere
else.

Issue #292 Label all required fields