Allow 9-digit zip code for personal care assistant
Update the form presented to the individual provider type "Personal Care
Assistant" to allow for 9-digit zip codes (plus the separating dash).

The rest of the infrastructure is already in place for 9-digit zip
codes, and it is working elsewhere; only the form needed to be changed.

Issue #290 Add 4 extra digits to zipcode field(s)