Add Gradle task to build Sphinx user docs
We want to automatically build the docs on a regular basis, and have
it be part of our existing build process.

Move the doc sources into a subproject of psm-app, and use the Sphinx
Gradle plugin[1] to build it. To view available formats:

```
./gradlew sphinx-userdocs:tasks
```

Co-authored with Sumana Harihareswara <sh@changeset.nyc>.

[1] https://github.com/trustin/sphinx-gradle-plugin

Issue #381 Add user help documentation building to Gradle build