Do not allow system or agent to see enrollments
Remove the system administrator (who can create and edit users) and the
service agent (whose role is currently being reconsidered) from the list
of roles with full access to enrollment data. Viewing the list of
enrollments as a user with either role should show an empty list, and
trying to view an existing enrollment (such as by going to a known URL)
will show an error to the user and log an "Access Denied" exception.

Note that visiting https://localhost:8443/cms/ops/viewDashboard will
still show recent events, which are things like "NPI 1234567893 was
approved". The links in these events do not work. We'll need to consider
if and how we want the recent events panel to work for the various
roles.

Issue #10 Limit sys admins', service agents' ability to view, edit
          enrollments