Explicitly name enrollment search path variable
The `@PathVariable` annotation can reflectively determine the name of
its variable when compiled with debugging symbols, but when compiled
without debugging symbols, attempting to load the page causes it to
throw an IllegalArgumentException with this error message:

    Name for argument type [java.lang.String] not available, and
    parameter name information not found in class file either.

Spring 3.2 either causes us to compile this file without the necessary
symbols, or has trouble reflecting. Either way, the solution is simple
and explicit, so even if later versions fix this we needn't come back to
this.

See also:
- https://objectpartners.com/2010/08/12/spring-pathvariable-head-slapper/
- https://stackoverflow.com/a/25797744/25637

Issue #219 Upgrade Spring