Upgrade spring-ldap to 2.0.2
This corresponds to spring-core 3.2.8.

The LDAP code is effectively dead at the moment, as we haven't spent any
time verifying that it works, documenting how to configure it, or
testing it. So, this is a blind version upgrade; the code compiles and
the application deploys, but that's all the more testing we can
realistically do right now.

Issue #219 Upgrade Spring