Save username of last login attempt in custom key
The spring-security-provided key is deprecated and will be removed.
Instead, use our own custom key to store the username from the last
attempted login, and prepopulate the login form with that username.

Note that we save the username in a custom handler that was built to
accomodate the MN-ITS system; although that system is disabled, we still
have a lot of code referring to it.

Issue #219 Upgrade Spring