Use Java K&R brace style
Opening braces should be on the same line as the statement that expects
a block, rather than alone on a line, and closing braces should be alone
on a line, except for linked control blocks (if/else, try/catch/finally,
do/while).

See also the Google style guide[1].

Fix the places we were not following this, and enforce it with
checkstyle.

Issue #456 Use consistent code style

[1] http://checkstyle.sourceforge.net/reports/google-java-style-20170228.html#s4.1.2-blocks-k-r-style