Add handlebars template variables to all requests
Create a new interceptor that adds the variables that handlebars
templates need. This allows us to stop manually adding these variables
via a call to a helper method; move the contents of that helper method
to the interceptor.

Also update the handlebars custom tag to include variables from all
scopes, rather than only PAGE_SCOPE, as adding variables in an
interceptor seems to add them to the REQUEST_SCOPE.

Remove the common variables added for handlebars templates (mainly
`ctx`), as they are now set once rather than needing to be set every
time a handlebars template is invoked.

Some of the variables are based on the principal user, an attribute of a
request set by spring security. Since we previously were only accessing
the principal on logged-in pages, we now need to handle the case of
anonymous access (allowed, for example, on the login page). Update the
`getPrincipal` helper to account for anonymous access.

Issue #238 Templatize UI