Enable CSRF protection
Cross Site Request Forgery "is a type of malicious exploit of a website
where unauthorized commands are transmitted from a user that the web
application trusts"[1].

Spring Security can protect against CSRF attacks[2][3] by including a
token on each page that must be submitted with a POST request, either as
a parameter or as a header.

Enable Spring Security's CSRF protection. This involves several pieces:

- Adding CSRF tokens to POST requests
  - <form:form>
  - other forms
  - Mustache
  - JavaScript
- Helping Spring Security find and parse CSRF tokens in multipart forms
- Parsing files
- Updating logout flow

---

- Adding CSRF tokens to POST requests

Adding CSRF tokens to forms is necessary for forms that are not
constructed using the Spring MVC helper tag `<form:form>`, which
automatically includes the CSRF token if needed. Forms in JSPs can use
the `<sec:csrfInput />` tag from Spring Security. Forms in
Mustache/Handlebars templates need to spell it out more completely.

JavaScript POST requests need to include the CSRF token in a header, and
in order to get that token we need to put it in the page. Add `<meta>`
elements to the `<head>` of all the pages that JavaScript can find and
use.  This includes the login page, which also has a form; I think this
is not a security issue because CSRF tokens are per-session, so an
attacker fetching the login page would get a different token than I
would get.

The CSRF token `<meta>` elements are in Mustache/Handlebars templates,
as are some of the forms themselves. Add the variable to the request
scope of every request using the newly-introduced interceptor.

- Helping Spring Security find and parse CSRF tokens in multipart forms

We need to do some extra work to get multipart forms working, which most
of the application flow uses. We're using the Apache Commons
FileUpload[4] package to help us parse files, and it processes all
multipart forms. In order for Spring Security to extract the CSRF from
such POST requests, the `CommonsMultipartResolver` must be named
`filterMultipartResolver`[5]

- Parsing files

When CSRF protection is enabled, Spring Security wraps requests in
classes like `SecurityContextHolderAwareRequestWrapper`. There is still
a `MultipartHttpServletRequest` in there somewhere, thanks to
FileUpload, but we need to dig a few layers to get to it.

- Updating logout flow

The logout action is no longer just a GET request, as that would allow
attackers to log out users. As a POST request, logging out needs a CSRF
token like any other POST request. Replace the "Logout" link at the top
of the page with a button styled like a link.

[1] https://en.wikipedia.org/wiki/Cross-site_request_forgery
[2] https://docs.spring.io/spring-security/site/docs/3.2.x/reference/htmlsingle/html5/#csrf
[3] https://docs.spring.io/spring-security/site/docs/current/reference/html/csrf.html
[4] https://commons.apache.org/proper/commons-fileupload/index.html
[5] https://stackoverflow.com/a/21448087/25637

Issue #219 Upgrade Spring