Upgrade Spring Security to 4.0.0
As the migration guide[1] notes, several of the default values changed.
Relevant to us are `use-expressions`, which I expect we'll revisit as
part of the authorization work we'll be doing; and the default names in
`form-login`. We previously updated the `logout` URL as part of the CSRF
work, so we are unaffected by that change, and of course we enabled CSRF
protection early in preparation for this change.

Several of the Spring Security beans altered their initialization from
injected settings to constructor arguments; update our configuration to
match.

[1] https://docs.spring.io/spring-security/site/migrate/current/3-to-4/html5/migrate-3-to-4-xml.html

Issue #219 Upgrade Spring