Use interceptor-provided current user variable
Instead of using the Spring Security helper tags, which will be
unavailable once we finish our template migration, use the variable
provided by the interceptor introduced in
31405615e8b1fcdbf3546cafaf03dfceeebaf9d6.

This also fixes an issue with errors on logged-out pages, as the error
page was not handling the case where the current user is an anonymous
user. This shows up as:

    The class 'java.lang.String' does not have the property 'enabled'.

The helper method used to fetch the current user correctly accounts for
anonymous users.

Issue #238 Templatize UI
Issue #504 Stack trace when mail server unavailable