Use PBKDF2 to hash user passwords
Replace the password hashing algorithm for users whose password is
stored in the database (as opposed to some external system, which is not
currently implemented). Previously we were using SHA1, which is not
appropriate for passwords; it is designed for speed, whereas password
hashing algorithms should be as slow as we can tolerate.

Use PBKDF2 instead, as provided by Spring Security, and expose the
tunable parameters (secret, iterations, and hash width) through
`cms.properties`. Hash width probably won't need to change, but users
should set the `secret` to something unique and the number of iterations
high enough that it takes their server about half a second to generate a
password.

See also:
- https://docs.spring.io/spring-security/site/docs/4.2.3.RELEASE/reference/html/crypto.html#spring-security-crypto-passwordencoders
- https://docs.spring.io/spring-security/site/docs/4.2.3.RELEASE/apidocs/

Issue #34 Improve password handling
Issue #465 Use FIPS-approved encryption