Remove trace logging
As per discussion in #16 and #91, we are going to be removing this
level of logging for now in a more general effort to clean up our
logging frameworks.  The signature for each method was a log feature
that is no longer relevant.

Issue #16 Manage sets of dependencies via Gradle or another tool
Issue #91 Fix "failed to lazily initialize" warnings in server log