Fix missing "change password" link for providers
The "change password" link is hidden behind a check that the user is an
internal user, as the PSM is not able to change the password for, eg,
users that signed in via LDAP. The variable that was set was
accidentally removed but not replaced in commit
801ef005aa1a2b7b164785b7cdd480291e1aec50 as part of the switch to
handlebars.

Set the `isInternalUser` variable in the interceptor, so that it is
available to JSPs and Handlebars templates alike. (It is currently only
used in one place.)

Issue #238 Templatize UI
PR #538 Use handlebars header, footer, logo, and nav templates in JSP
        templates