Use links in table rows on `My Profile` page
Rather than radio buttons and a form with two submit buttons.
This design is more in line with similar tables in the rest of
the PSM and, by removing the form, avoids the issue of having
accessible submit buttons for the form.

Issue #553 Use accessible submit buttons on all forms