Fix check for pending submissions
When we restricted the system administrator role by removing it from the
list of roles that have full access in commit
83f2cbc22886e4c8ab856e0b6fc3ce42e800f1ce, the check for pending
submissions broke. This check is designed to prevent multiple enrollment
submissions, as after the first no more will be accepted, and instead an
error message will be displayed:

    There is a pending request for the same profile. Submission will not
    be allowed unless the other request is denied.

The check broke because it was searching for pending enrollments using
the built-in system user, which previously caused the search function
not to enforce user privileges. This check was the only use of that
search function with the system user, so replace the call with a greatly
simplified version of the search function.

Issue #546 Post processing fails on submitted enrollment applications