Test that change password link exists: #568
PR #568 fixed an issue where the change password link wasn't showing up
for providers.  This tests that, and will fail if the link disappears
again.  This also adds ignored features to be implemented later, testing
whether the change password link functions correctly.  This uses the
general steps library added for the logout test