Use per-instance, not static, loggers
We've settled on a pattern of having each instance of a class that needs
to log data have a non-static instance of a `Logger`. Update the classes
that used static loggers to instead use non-static loggers, and update
the variable name to match.

This might also ease the WildFly out-of-memory problems, if they are in
fact caused by static instances hanging around across deploys, although
I haven't confirmed that theory.

Issue #16 Manage sets of dependencies via Gradle or another tool
Issue #421 Add Wildfly restart to CD script