Fix login on logged-out error page
The page shown when an error happens without a signed-in user (such as
trying to self-register when the application server cannot reach the
email server) contains a login form. However, the target of the form was
a relative path, which meant that, in cases such as the
self-registration error, where the URL has additional path components,
the browser would try to send the form data to that relative URL - e.g.,
`/accounts/login`, which results in a 404.

Specify the full path of `/login`, which matches the path set in
spring-security.xml.