Change license state in test: #571
This is a temporary fix.  Because of a MN-specific rule, enrollments for
some provider types cannot be approved unless they have a license issued
by the state of Minnesota.  Eventually, we need to remove the rule.  For
now, this change will allow us to approve enrollment applications
generated by the integration tests.