Improve color contrast for links on light blue table rows
For accessibility, we need sufficient color contrast for links on
table rows with a light blue background.  This commit darkens the
previous link color from #1aabff (HSL: 202, 100%, 55.1%)
to #0071b3 (HSL: 202, 100%, 35.1%).  It also removes redundant
css rules that apply this color where it is already applied.

For links on dark blue backgrounds (like the `Help` and `Logout`
links in the header) this commit restores the dark blue
background to its original color (#0d4478) (it was previously
lightened in commit 799d05b808cec13f61604999d9dad2e02015c98e)
and lightens the links from #1aabff to #43baff
(HSL: 202, 100%, 63.1%).

The result is two blue link colors, #007b3 for most links, and a
lighter #43baff for links on dark backgrounds.

Issue #465 Improve ADA compliance