Allow system role full access for processing
When I removed access from the system role in
83f2cbc22886e4c8ab856e0b6fc3ce42e800f1ce, we didn't understand how
pervasive the PSM used the system role for internal processing. The
first visible problem was that the post-processing of a submitted
enrollment application failed, as the post-processing code no longer had
permission to read or modify the application, but this was not the only
problem.

We will continue to change internal processing code to no longer require
the system user to have full access, but in the meantime, restore the
functionality of the system to unblock development of other issues.

This partially reverts commit 83f2cbc22886e4c8ab856e0b6fc3ce42e800f1ce.

Issue #10 Limit sys admins', service agents' ability to view, edit
          enrollments
Issue #546 Post processing fails on submitted enrollment applications
PR #524 Restrict agents and system administrators