Bundle generated user help docs in web app
Build and include the HTML user documentation in the built PSM web
application, and link to it from the (previously non-functional) "Help"
link at the top of every page.

In order for `cms-web` to know about the output of `:userdocs:html`, we
need to declare a configuration, and then link the output of the task to
the configuration. For now, as we only need the HTML output, I only
declared that configuration and output; hopefully this will be a useful
template in case we ever need to refer to the results of the other
tasks.

See also:
https://docs.gradle.org/current/userguide/artifact_management.html
https://mrhaki.blogspot.com/2016/02/gradle-goodness-inter-project-artifact.html
https://discuss.gradle.org/t/how-to-copy-files-from-a-custom-configuration-into-a-particular-war-directory/6004

Issue #154 "Help" link does not display help
Issue #359 Make user help docs reliably user-visible