Remove unused libraries
When we transitioned to gradle, there were some libraries that we were
depending on for reasons we didn't understand, but we separated
reviewing the dependencies from the conversion to gradle.

The last few libraries are not, as far as I can tell, ever used. They
are not referenced anywhere in our project directly, and they are not
referenced by any of the TopCoder libraries. They're not even depended
on by any of our dependencies.

Additionally, `jackson-mapper-asl:1.9.13` is provided by WildFly (both
10 and 11), so even if we did need it, we need not bundle it into our
application.

Issue #16 Manage sets of dependencies via Gradle or another tool