Whitespace and empty classes
This is a whitespace commit made as part of a larger set of changes
related to accessibility (as described in issue #554).  There were
also a few cases where divs had empty class attributes, which I
removed just in the spirit of cleanining as I worked.

Issue #554 Give form input elements a name for accessibility