Defined form prefix variables
In order to provide unique identifiers on several forms in an effort
to make forms and labels align, a variable is now being set on the
top of form-heavy jsps which can be used as a prefix for DOM level
identifiers.

Note that for some particularly heavy form pages this ID should
actually be changed through the page, so that elements within a
single jsp don't clash.

Issue #554 Give form input elements a name for accessibility