Added titles to form elements
In order to make fields accessible to screen readers we need
all form elements to either have a title, or to have a label
associated using a for element.  In many cases labels exist,
but in some cases there is no label or the label applies to
more than one field.  In these instances a title has been given
to the form element instead.

Issue #554 Give form input elements a name for accessibility