Label and field associations
This commit is rather large, it associates labels with form elements
in all cases possible using for and id attributes on the related
DOM elements.

Many of the forms used a pattern of formName to name form elements.
This name was combined with a form level ID to attempt to ensure unique
ids.  In order to use this formName in a label association the variable
definitions needed to be moved above the labels; unfortunately git did
not realize that's what was happening (it logged the moving DOWN of the
other html elements to below the variable definition).  This means it
isn't really possible to separate out that part of the change in a
clean way.

It is possible this commit results in id conflicts.

Issue #554 Give form input elements a name for accessibility