Reproducibly build API docs by omitting timestamp
From the Reproducible Builds project[1]:

    Often the time of the build was used as an approximate way to know
    which version of the source has been built, and which tools had been
    used to do it. With reproducible builds, recording the time of the
    build becomes meaningless: on one side, the source code needs to be
    tracked more accurately than just a timestamp, and on the other
    side, the build environment needs to be defined or extensively
    recorded.

Tell Javadoc not to include the timestamp in the API documentation to
allow for reproducible builds. This cuts down on noise in diffs, which
is particularly useful while we're using GitHub Pages to host these
docs, as it prevents needless commits.

PR #478 De-automate Javadocs publication to GitHub Pages

[1] https://reproducible-builds.org/docs/timestamps/