Remove jasypt library and undocumented encryption
The PSM previously allowed for encrypting values in `cms.properties`,
but this functionality was barely documented - only a single mention in
the header with a link to a page on the jasypt site.

The jasypt library, itself, seems to be abandoned - the last release was
in February 2014. It's also not entirely clear that its
encryption/decryption routines are robust[1].

If this functionality turns out to be worthwhile, we should reimplement
it using modern, maintained, secure alternatives to jasypt.

Issue #16 Manage sets of dependencies via Gradle or another tool

[1] https://security.stackexchange.com/a/65240