WIP add Spring cache-busting
Configure Spring to refer to static assets by URLs that include a hash
of their contents, so that web browsers can aggressively cache but not
end up with stale CSS and JavaScript.

WIP because most of our references are either in Handlebars templates,
which are not affected, or because they're not using the `<c:url>` or
`<spring:url>` in our JSPs.

https://spring.io/blog/2014/07/24/spring-framework-4-1-handling-static-web-resources

Issue #622 Prevent issues from browsers caching outdated CSS