Explicitly depend on Hibernate to compile
WildFly ships with several libraries, which are available to
applications deployed to it at runtime. However, we should not require
WildFly to be present in a particular directory to compile the PSM.

Move the Hibernate dependency from WildFly to gradle.

Issue #16 Manage sets of dependencies via Gradle or another tool