Explicitly depend on WildFly-provided libraries
WildFly ships with several libraries, which are available to
applications deployed to it at runtime. However, we should not require
WildFly to be present in a particular directory to compile the PSM.

Replace the various JBoss javax dependencies from WildFly with the
official, generic (Oracle) versions from Maven Central.

Issue #16 Manage sets of dependencies via Gradle or another tool