Remove rule and process around mailbox accounts
There was an unimplemented stage of our enrollment business process that
said that each enrollment application requires a mailbox account, and
then did nothing if they did not already have a mailbox account. While I
don't know for sure what a mailbox account is in this context, I can
tell we don't need to have it in this (barely) partially-implemented
state; delete it.

Issue #165 Remove/edit code pointing to old external-sources app