Move JavaScript code into its own subproject
Create a new `frontend` subproject in gradle, and wire it up to be a
dependency of `cms-web`, so that the code in that project is put into
the same directory at build time as it was previously. This is a
prerequisite step to start adding npm modules and having a build process
around our JavaScript code.

Issue #626 Manage JavaScript dependencies and modernize build