Add npm and nodejs
Use a gradle plugin to manage npm and nodejs, so that it will be the
same across different machines. Use the current latest version of both.
We're not yet using either, but this will prepare us to do so. It adds
the `frontend:npmSetup` task, which will install nodejs and npm into
`psm-app/frontend/.gradle/`, and `frontend:npmInstall` task, which -
once we have dependencies or dev dependencies in `package.json` - will
download dependencies into `psm-app/frontend/node_modules/`.

Set up a symlink to the installed version of `npm` in `frontend/` for
ease of use.

Issue #626 Manage JavaScript dependencies and modernize build