Add PsmPage, PsmStepDefinitions classes for accessibility tests
Since accessibility tests will happen for every page, use
a PsmPage class and a PsmStepDefinitions class to avoid needless
repetition and reduce boilerplate.

In the checkAccessibility method on the PsmPage class, we
set a "runOnly" option to check the page for accessibilty
against the "WCAG 2.0 Level AA" standard.  This standard
is a superset of the "WCAG 2.0 Level A" standard, but axe-core
treats them as separate sets of rules, so we use the tags for
both of them ("wcag2a" and "wcag2aa").  See the axe-core docs:
https://axe-core.org/docs/#options-parameter

Issue #518 Implement accessibility checking in CI