Remove GetTicketDetailsResponse JAXB data type
This XML-defined object was used only in a single place, and only to
wrap a single return type from a single method call. Unwrap that return
type and delete the data transfer object to make both the calling code
and the called code easier to read and easier to change.

Issue #596 Reduce use of JAXB class generation