Remove ResubmitTicketRequest JAXB data type
This XML-defined object was used only in a single place, and only to
wrap a few method arguments to a single method call. Unwrap the method
arguments and delete the data transfer object to make both the calling
code and the called code easier to read and easier to change.

Issue #596 Reduce use of JAXB class generation