Fix agreement documents for organization providers
The form for gathering the provider statement for an organization
provider type was broken when adding labels in
dce382547555218c53e24e32440b7f2a11e53269. The URL of the agreement
document, the name of the agreement document, and a hidden input each
depend on a variable set, in-order, in the JSP; by rearranging the
label, the link to the agreement document was broken.

Rearrange the agreement document variables and elements to match their
original order (but now within a `<label>`), and update the `<c:set>`
JSP tags to be empty elements (`<c:set ... />`) instead of start + end
tags (`<c:set ...></c:set>`).

Issue #554 Give form input elements a name for accessibility
PR #583 Associate labels and add titles