Remove PECOS verification
The PECOS jBPM work item handler was never referenced, either in our
business process or in the step-to-code mapping jBPM set up.

Since it depends on the legacy external data sources, which no longer
exist, and since it is not wired up, delete the handler, the API client
it depended on, the XSLT the client used, and the JAXB data model the
client used.

Also delete the associated checkbox from the admin review page, its
Drools rule, and its backing fields in the MVC, JAXB, Hibernate, and
PostgreSQL representations.

Fully removing all traces of the current, partially implemented PECOS
verification will make it easier to add it back in later, after we've
improved the data model, updated the UI, and set up access to PECOS
data.

Issue #165 Remove/edit code pointing to old external-sources app
Issue #596 Reduce use of JAXB class generation
Issue #761 Save screening results in the database