Add Spring Security rules for profile pages
When an admin user is on the profile page, logs out from another browser
tab, and then refreshes the profile page, a stack trace was displayed
instead of redirecting to the login page. The problem was caused by path
prefixes of the profile page not being listed in spring-security.xml.

Issue #431 Expired session error from admin profile page