Disable Travis CI
As our project has grown, the value in having our own build server
became more and more clear. We set up a Jenkins instance, and migrated
all of our jobs over. Now that the last job is migrated, we no longer
need Travis; delete its configuration, scripts, and our references to it
in our documentation.

Note that this also deletes the Sauce Labs configuration. If we later
decide to use Sauce Labs with Jenkins, or on our dev machines, we can
add it back in.

Issue #629 Use Jenkins for continuous integration