Move 'View' link from NPI number to Actions column in tables
This is on the provider dashboard and enrollments pages.
'Edit' and 'View' links do the same thing for draft
enrollments, so show 'edit' for drafts and 'view' for other
enrollments, in the Actions column. Adjust column widths
to keep Actions column text from wrapping to two lines.

Issue #672 Empty link text for providers with no NPI