Add FHIR API using hapi-fhir
Use the HAPI-FHIR library[1] to create a REST API[2]. The library takes
care of the FHIR housekeeping methods, and I've rooted the API at
`/fhir` - which means that, with the application prefix, the API can be
reached at `/cms/fhir`.

The FHIR housekeeping methods are defined by the FHIR HTTP
specification[3], and include the `/metadata` endpoint, which in turn
links to other endpoints. Again, with the root URI and the application
context, the full endpoint would be `/cms/fhir/metadata`.

Note that this does not yet add any functionality; it just lays a
foundation for us to build out an API.

[1] http://hapifhir.io/
[2] http://hapifhir.io/doc_rest_server.html
[3] https://www.hl7.org/fhir/http.html

Issue #641 Create API that returns a list of approved and/or rejected
           providers